fix: move botoes para pasta shared de componentes
