+++ src/views/EditNoteView.vue
@@ -1,5 +1,5 @@
 <script lang="ts">
-import PrimaryButton from '@/components/button/PrimaryButton.vue'
+import PrimaryButton from '@/components/shared/buttons/PrimaryButton.vue'
 import HeaderApp from '@/components/shared/HeaderApp.vue'
 import router from '@/router'
 import { toast, type ToastOptions } from 'vue3-toastify'
